FULL CODE BASE/requirements.txt
dspy-ai>=2.0.0
langchain>=0.1.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
jupyter>=1.0.0
tqdm>=4.64.0